modify kicker

+++ .Kicker/common_01.sh
@@ -11,13 +11,15 @@ echo "*********************************************"
 echo "step02: create inventory and properties..."
 echo "*********************************************"
 cp Env/i_${ROLE}.yml Env/inventory.yml
-cp Env/p_${ROLE}.yml Env/properties.yml
+if [ -e Env/p_${ROLE}.yml ] ; then
+  cp Env/p_${ROLE}.yml Env/properties.yml
+fi
 
 # Vagrant(pre)
 echo "*********************************************"
 echo "step03: launch target vms..."
 echo "*********************************************"
-cd Vagrant/${VM01}
+cd .Vagrant/${VM01}
 vagrant up
 
 # Itamae
@@ -25,27 +27,27 @@ echo "*********************************************"
 echo "step04: deploying by itamae..."
 echo "*********************************************"
 cd ../..
-Itamae/itamae_exec.sh
+Bin/suit itamae exec
 
 # Serverspec
 echo "*********************************************"
 echo "step05: test by serverspec..."
 echo "*********************************************"
-Serverspec/serverspec_exec.sh
+Bin/suit serversepc exec
 RC_SS=$?
 
 # Infrataster
 echo "*********************************************"
 echo "step06: test by infrataster..."
 echo "*********************************************"
-Infrataster/infrataster_exec.sh
+Bin/suit infrataster exec
 RC_IT=$?
 
 # Vagrant(post)
 echo "*********************************************"
 echo "step07: delete target vms..."
 echo "*********************************************"
-cd Vagrant/${VM01}
+cd .Vagrant/${VM01}
 vagrant destroy -f
 
 # represent exit status
